feat: Utilize 'explain' RPC methods
ChatSearch does not call the web service endpoint
any more. RPC provided by the appmap package
provides everything that's needed for the UI.
