+++ packages/cli/tests/integration/RpcResponse.ts
@@ -1,0 +1,6 @@
+export type RpcResponse<T> = {
+  error?: { code: number; message: string };
+  id: number;
+  jsonrpc: string;
+  result?: T;
+};
