+++ packages/cli/tests/unit/fixtures/ruby/app/models/api_key.rb
@@ -1,0 +1,1 @@
+# dummy
